Refactor asserts in DataTestSubscriber
Prior to this commit, the DataTestSubscriber would only have a
`assertNextSignals(T... signals)` which checks for
*strict object equality* between the Next received signals
and the provided arguments.

This commit renames that method to `assertNextSignalsEqual` and adds a
new one `assertNextSignals(Consumer<T>... expectations)`, which apply
each Next signal to the expectations given as arguments. This allows new
ways to verify the received Next signals, in a more flexible way.