Expose Tuple constructors
Stream combinators, mappers and other constructs used to manipulate
streams can be extracted to standalone functions. It is desirable to
test these functions in isolation, i.e. without initializing a Reactor
environment.

Unfortunately the Tuple constructors are package private, making it hard
to create instances for testing purposes. This changes makes them puplic
for the sole purposes of testability in projects that use the reactor
project.