repeatWhenEmpty Iterations
Previously, a fix when into the repeatWhenEmpty implementations to deal with
finite versus infinite repeat loops.  This fix broke two pieces of
functionality.  First, it returned the wrong error when the max retries was
exceeded.  Second it returned 0 for each iteration of the loop, rather than
incrementing as designed.  This broke any functionality dependent on the
iteration counter (like a backoff algorithm).

This change refactors the implementation to fix both of those problems and
updates the tests to more accurately test that both bits of functionality are
satisfied.