Tuple Hash Code
Previously, Tuple hash codes were broken because at the root of the hierarchy,
Object identity was used to start the calculation.  This bug was due to the
recent removal of the root Tuple type, leaving Tuple2 calling
Object.hashCode() instead of a properly structured Tuple.hashCode().  This
change fixes the bug by using the t1 as the start of the hash code
calculation.