fix #196 Add subscribe with Consumer<Subscription>
This change adds the possibility to provide a Consumer of Subscription
when subscribing to a Flux or Mono. The Flux/Mono variants of subscribe
that allows this are the most complete ones (all lambdas are provided
and the prefetch is set).

The underlying LambdaSubscriber and LambdaFirstSubscriber are built
with this additional Consumer, which can be null if not relevant.

The subscriber is protected against errors during the consumption of
the Subscription, but such an error will prevent the call to request
on said Subscription.

Testing for LambdaSubscriber's onSubscription behavior has been added.