fix #178 Add mergeSequential and flatMapSequential operators
This commit adds mergeSequential and flatMapSequential operators to Flux
which allow to merge sequences without interleaving them. Compared to
concat/concatMap, the benefit lies in the fact that these operators
subscribe to the inner sequences eagerly (eg. allowing better
parallelism).

A variation of merge that exposes the delayError parameter has also been
added.