see #238 Revert Flux.then(Mono<V>) introduction
This commit partially reverts commit dbc8ba2 by removing the Flux method
`then(Mono<V>)`. This is because said method can clash with
`then(Publisher<Void>)` when type of V is Void.

The clashing method `then(Publisher<Void>)` has been marked as
deprecated and a thenEmpty alias created, so that the removed method can
be reintroduced in 3.1.