Add logging and aliasing capacity to StepVerifier
This commit adds the notion of a description to all StepVerifier events.
Most expectations have a relevant description to begin with, but the
`as` operator also allows the user to choose a different alias for the
previous expectation in the chain. When producing an AssertionError
through the `fail` method, this description is used at the beginning of
the message.

Second feature added is the `log()` method that can be called just
before `verify`. It activates a logger in the StepVerifier that will
notably list the descriptions of every task in the verification scenario
as well as some execution details (for now aliasing during conflation).