fix #254 Add BaseSubscriber to write subscribers with request
When writing a simple Subscriber, use of lambda-based overrides is good
and simple enough. However, as soon as a library user needs to issue
request(n) manually, things get a little bloated: an anonymous instance
needs to be written, the Subscription captured and reused in onNext to
perform request.

This commit offers an intermediate option, BaseSubscriber, that captures
the Subscription and exposes the request method.
It also contains hooks for the Subscriber methods (as the actual
implementations of these methods include boilerplate code). These are
prefixed with `hookOn`.