fix #257 Don't silence errorNotImplemented in Schedulers
This commit fixes the fact that Scheduler workers can silence 2 classes
of Exceptions specific to Reactor: BubbleException and
OnErrorNotImplemented. This is due to Exceptions.throwIfFatal throwing
these, which in turn prevents Schedulers.handleError to forward them to
the Thread uncaught exception handlers.

The throwIfFatal has been split, introducing a throwIfJvmFatal that only
deals with JVM-specific fatal errors. This new method is used to handle
Schedulers fatal errors.

Additionally, a hook has been introduced that will be executed if
Schedulers.handleError actually handles the errors (ie it isn't thrown).