fix #262 Prevent log() of Subscription as a Collection
why
---
When log() is used, the Subscription itself is logged. Some logger
implementations (eg. Log4j) attempt to detect a Collection in the list
of arguments it logs, in order to apply special logic (log the string
representation of all elements using the iterator).
Since some Subscription implementations partially implement Collection
(usually a Queue), but without implementing `iterator()`, these fail.

modifications
-------------
This commit makes SignalLogger log the string representation of the
Subscription, so that underlying loggers won't wrongfully attempt to log
it as a Collection.

Additionally, SignalLogger can be instantiated with a Logger supplier,
mainly used for tests.