Use compileOnly configuration for JSR-166 backport dependency (#292)
compileOnly better expresses how the dependency is used and that it's
expected to be provided by the JDK (when running on Java 9). It also
allows the Spring IO Platform checks to pass as it no longer
identifies are dependency that it not managed by the Platform.

The use of compileOnly requires upgrading to Gradle 2.12 or later.
This commit upgrades to 2.14.1, the latest 2.x release at the time of
writing.