fix #199 best effort WorkQueueProcessor limit detection
This commit attempts to avoid letting too many subscribers subscribe to
a WorkQueueProcessor if the underlying executor's limit can be detected.

It does so in a best effort fashion, introspecting the ExecutorService
implementations that expose a way of getting that information (namely
ThreadPoolExecutor and ForkJoinPool). For others, it is still the user's
responsibility of avoiding too many subscribers.

The introspection is done in a testable static method that returns
Integer.MIN_VALUE if no information can be obtained.