fix #264 Add Flux.doOnEach with Signal consumer
This commit adds a doOnEach method that converts onNext, onError and
onComplete signals to a Signal instance before passing it to a
unique consumer.

As an advanced use case, this can also be used to pass the events to
a class that implements Subscriber for event forwarding. Note however
that this is discouraged, as the class has to be very careful of
its state and concurrent behavior. eg. what if it is actually
used in a .subscribe in addition to receiving these forwarded events?