fix #289 Split Signal into immutable and mutable version
Signal is now an abstract class, and the default implementation is still
the immutable one. Introduce a mutable version for onNext signals that
can be used to limit the number of created instances in advanced cases,
here used for doOnEach.

A new operator, `FluxPeekStateful`, is used. It peeks into a sequence
while also passing around an arbitrary state object, per subscriber.

Note that FluxPeekStateful already doesn't call onError upon an
afterTerminate callback failure (#270).