fix #301 flatMapSequential broken with more than prefetch src elements
The logic in creating the bounded queue for inner subscribers is
inverted and thus takes the minimum of maxConcurrency and prefetch,
undersizing the queue.

Unfortunately, the queue offering of new subscribers isn't checked, and
thus as soon as there are more than prefetch subscribers, the excess is
silently discarded :(

This commit fixes the situation by using the correct "max" logic for
queue sizing.

Additionally it adds a sanity check when doing the eager inner
subscribe, and allow testing of that logic by passing a custom queue
supplier.

Reworded javadoc to highlight the difference with concat/concatMap and
merge/flatMap.