rework #274 Consistent scheduler rejection/shutdown, Worker.isShutdown
This commit reworks #274 by more consistently checking the REJECTED
result of scheduling so it can be transformer to a
RejectedExecutionException.

That RejectedExecutionException is often thrown, but is passed to
onError instead when it is possible. If it happens during onNext, the
subscription is cancelled. In request, the REJECTED state is ignored as
it cannot be taken into account (no possibility of throwing or calling
onError).

The shutdown of Schedulers has been mutualized in a method that attempt
to make it safer.

Lastly, the `isShutdown` method has been added to Worker. This allows
FluxPublishOn to check that state in onComplete, preventing it to
wrongly throw the RejectedExecutionException.