fix #338 Deprecate zip(Iterable), fix bug with size < 8
Flux.zip(Iterable) is a clunky API: as it combines to Tuples, it limits
the number of acceptable sources to 1-8. Furthermore, the Tuple2 return
type means that you either need to know the number of sources in order
to cast to the relevant TupleN (in which case you should instead use
zip(source1, source2, ...)) OR you can only use the tuple as Iterable.

In all cases, combining to a List would be more relevant, and usage of
a custom combinator or of the fixed sources alternatives is to be
preferred from now on.

Using a custom combinator like Arrays::asList is susceptible to a bug
if the number of sources is below 8 though: the scalars Object array is
initialized to a length of 8 in any case, and grows beyond if needed,
but it is not trimmed in FluxZip#handleIterableMode if we have less than
8 sources. This is now fixed and tested.