fix #354 Add decorator hook for ExecutorServices in Schedulers
This can be useful for transverse instrumentation, or usages where one
wants to replace the underlying ExecutorService used by the default
Schedulers, without changing the Schedulers themselves.

The idea is that each Scheduler implementation that uses an
ExecutorService would instantiate it but also pass it through the hook
in Schedulers. Same for ScheduledExecutorService.