keep the stacktrace of onExpectation failures (suppressed)
When an Exception that is not an AssertionError occurs during the
onExpectations phase, it is wrapped in an AssertionError. Problem is
that only the exception name and message are kept, but no the stacktrace
of the original exception. By suppressing it, we ensure the stacktrace
is still available to the end user for debugging purposes.