Fix ParallelFlux getPrefetch not always accurate + minor fixes
operators with a notion of prefetch all report their parameter instead
of the default -1. Peek and hide both will report the source's prefetch
as well as the source's isOrdered.

creating a ParallelFlux with from now correctly report if a source
is null (including the vararg version receiving a null as array).

extracted the sortedMerger function so it can be more easily tested.