Tuples toString format change (nulls, brackets)
This commit changes the toString representation of Tuple2 through
Tuple8. The new format is enclosed in square brackets, and includes
a comma separator between ALL values.

Sparse tuples (with nulls) have their null values represented as an
empty string but a separator is now also included.

For example `Tuple3(null, 2, 3)`` is "[,2,3]", and an empty Tuple3
would be "[,,]".