fix #363 Deprecate and alias ambiguous window/buffer (size or time)
This commit deprecates the buffer/bufferMillis and window/windowMillis
signatures that deal with max size plus a timeout. As they use
`int, long` parameters, they are ambiguous with the "timespan and
timeshift" versions, which use `long, long`.

So these variants have been renamed with the `bufferTimeout` prefix
instead, and the old signatures have been deprecated.