fix #417 Add checkpoint() operator to force single assembly tracing
This commit adds a new `checkpoint` operator that allows to
assembly-trace a particular Flux or Mono.

Additionally, one can specify a `description` and/or `correlationId`
that will both appear in the trace.

The usual path to assembly tracing (the debug mode hook) doesn't define
such a description nor id.

TODO: add operator to Flux derivatives, javadoc