fix 88 StepVerifier's VirtualTimeScheduler activates for all by default
This commit ensures the VirtualTimeScheduler created by StepVerifier
is activated on all schedulers of the SchedulerFactory by default,
instead of just the `timer()` previously. This is in preparation of a
change in core where all these Schedulers are time-capable and some
operator would use eg. parallel() as a default scheduler instead of
timer().

This can have side effects though, eg. when testing backpressure with
expectNoEvent and thenRequest, StepVerifier can run forever (if you're
not using the Duration variant). Adding a thenAwait after the request
will work around this issue.

(fix reactor/reactor-addons#88)