fix #480 deprecate boolean delayError signatures, replace with suffix
This commit deprecates most Flux operators with a boolean delayError, in
favor of replacing these with a simple method defaulting to false (if it
doesn't exist) and a xxxDelayError variant defaulting to true.

The only exception is publishOn, which already default delayError to
true (so removing the boolean signature and exposing a suffixed method
would imply that default changes).