fix #516 Rename Mono.flatMap to flatMapMany, then(Function) now flatMap
This commit changes the meaning of flatMap on a Mono. It now returns a
Mono as well, obtained by applying the flatMap function to the source's
value. This was previously the behavior of Mono.then(Function), which
has been renamed to flatMap.

As a consequence, Mono.flatMap has been renamed to Mono.flatMapMany to
better convey that it maps to a multi-value Publisher and flattens to a
Flux<R>.