fix #555 Rework FAQ on then/and after `then` renaming
The FAQ entry attempted to make more explicit that the then signature
that used to take a Function was to be handled with care when used
after an empty or Void sequence. This is no longer necessary for then
since that variant has been renamed to `flatMap`, making it all the
more explicit.