fix #566 Simplify Mono.subscribe, move the old behavior in toProcessor
This commit simplifies the `subscribe()` method of Mono to align it with
Flux, returning a Disposable (that is actually a no-callback subscriber
similar to the one returned in callback-based subscribe methods).

If the current instance is a MonoProcessor, connect that MonoProcessor
like before.

The wrapping into a returned MonoProcessor is now accessible in the
`toProcessor` method. However to exactly reproduce the previous behavior
one will need to call .toProcessor() then .subscribe(). This won't
always be hinted at by a compilation error.