fix #587 Checkpoint with identifier is light, skips stacktrace
This commit introduces a lighter version of checkpoint() that avoids
the filling of a stacktrace when a String identifier is provided. Users
can still opt-out of this optimization by using the
`checkpoint(String, boolean)` variant instead of the light
`checkpoint(String)` one. The no-arg version is still based on using a
stacktrace.

JMH benchmark shows an improvement of an order of magnitude in
throughput.