fix #484 Fix windowTimeout timeout starting behavior
This commit changes the behavior of windowTimeout so that it starts its
timer right from the subscription. Subsequent timers are started
whenever a window closes (either from timeout or from size).

Additionally, this has the effect of emitting empty windows, even in the
case where multiple timeouts occurred. Thus, this operator can be used
for heartbeat kind of scenarios, where a new empty window will complete
every "maxDuration".

In order to not impact the bufferTimeout operator, the operator classes
don't share a common base class anymore.