fix #616 Keep one significant factory method for Processors with Builder
This commit removes the recent deprecation on a few factory methods for
Processors that now have a Builder, keeping most significant and short
constructions from the boilerplate of using a builder.

WorkQueueProcessor and TopicProcessor keep the create/share with name
and bufferSize, UnicastProcessor keeps the queue+terminationCallback
variant and EmitterProcessor keeps the bufferSize variant.