fix JMH IntelliJ integration: back to plugin, not imported as src scope
This goes back to the champeau.jmh plugin, but ensures that 1) it is
only applied to core module and 2) IntelliJ doesn't mark the src/jmh
folder as a source folder (thus preventing it to compile). The later is
done by using the idea plugin.

The [check] task doesnt't run the jmh bench, only [jmh] does.