fix #614 Null analysis enabled through the use of Nullable/NonNullApi
This commit enables null-safety analysis by IDE like IntelliJ and other
tooling through the use of JSR 305 annotations. Although the JSR is
dormant, for this purpose it provides a common ground that can be
exploited by tools to check for null misuse.

The JSR's `Nullable` annotation is directly used, and we introduce a
`NonNullApi` that is based on JSR305 meta-annotations.

`NonNullApi` is put at package level on all packages to specify that
Reactor API is non-null most of the time. The exceptions, either in
return values or input parameters, are annotated with `Nullable`.

This commit covers most of the ground, but might be followed by further
work as problems are detected following the introduction of the
annotations.