Attempt to test only reactor-test if all changes are in `reactor-test/`
This works by overriding the default build script from Travis CI and
delegating to a bash script (so that any errors will correctly fail the
build).

The script detects a PR build vs a push build, and looks for any
affected file not starting with `reactor-test/`:

 - PR build: lists the files changed by the commits in the whole PR
   (FETCH_HEAD..TRAVIS_BRANCH, the later being the target branch of the
   PR)
 - push build: use the commit range provided by Travis
   (TRAVIS_COMMIT_RANGE) but work around a few issues. The commit range
   from travis uses `...` notation and we fix it to `..` (as noted in
   issue travis-ci#4596). Furthermore, the range is invalid in case of
   a brand new PR or a rebase/force push. In this case, the diff fails
   and we fallback to only looking at the last commit. This is not ideal
   but the associated PR build should at least test the whole code.

The grep is inverted (-v) and quiet (-q), meaning it will succeed if
anything else than reactor-test is present.

master and 3.0.x push builds are detected in order to force a full test,
as these are merge commits.