fix #661 Macro-fusion of flux then and mono then use incompatible T[]
This commit fixes a bug around macro-fusion of `then(Mono<T>)` when
mixing a Flux suppressed source and a Mono suppressed source (see #547).

This is due to the Flux's usage of `then` resulting in a Publisher array
of sources whereas the Mono usage still results in a Mono array. When
attempting the two, the type difference causes an ArrayStoreException.