fix #664 Propagate failing callback from doOnNext to doOnError if fused
This commit resolves the situation when a doOnNext and a doOnError are
chained, there is async fusion and the first callback throws an
exception.

This would result in the `doOnError` never being invoked, due to its
`poll` invoking the `doOnNext.poll` without a try/catch.

Now the inner poll call is protected in a try/catch block and if it
fails, the doOnError callback is invoked. Additionally, the exception
will be _propagated_ (and the sequence most probably stopped), and the
case where the error callback itself also fails is covered as well.