fix #670 Add expect|verifyErrorSatisfies StepVerifier error expectations
This commit adds two convenience methods: expectErrorSatisfies and
verifyErrorSatisfies, taking a Consumer that is expected to apply
assertions to the error.

This is similar to consumeErrorWith, but more discoverable (due to
naming alignment) and with the slight difference that the consumer's
assertion error message is wrapped inside a StepVerifier-specific
AssertionError.