fix #690 Null out source in MonoProcessor upon all terminations
This commit nulls out the reference to the source whenever the processor
terminates. This helps preventing retention of a source and potentially
data that is only associated with it, while still holding on the final
data in the MonoProcessor.

For example, having a Future as the source, and letting the processor
cache a mapped version of the future's data would still retain both the
Future, its value and the mapped value before this change. Now only the
map value is tied to the lifecycle of the processor.