fix #716 Add error scalar support
This commit makes operators that had a path consistently ending in
onError Callable. In turn, the `call` method will throw an exception for
these cases.

As this makes the makes the error synchronously available (at the slight
cost of a try-catch at assembly time), this may change the behavior a
bit when combined with operators that delay errors, which won't have the
opportunity to do so unless the ScalarCallable is followed by `hide()`.