fix #724 Flux.sample() includes last source element as if sampled
This commit ensures that the last element in the source sequence is
emitted in the case the last sampling window is not closed before the
source terminates.

Consider a source that emits A B C D E at 100ms intervals and a sampling
timespan of 350ms: [--A--B--C-][-D--E-|)->

The part between [] is a first full sample over 350ms, containing A B
and C, and thus resulting in C being emitted. The second part between [)
is NOT a full sample as only 200ms elapse before onComplete.

Nevertheless, with this change the [) part is considered a sample and E
is also emitted.