fix #771 Exceptions.addThrowable now reuse same root composite
This commit fixes a bug where `addThrowable` would keep creating nested
composite exceptions for Throwables past the second one, resulting in a
deeply nested structure with composites of 1 composite + 1 legit error.

This is now fixed, and past the first error a single CompositeException
is created and reused for further errors, which all end up as suppressed
siblings.