fix #785 scanOrDefault ignores the Attr default
This commit makes scanOrDefault more intuitive to use: it will always
fall back to the user-provided default if the Scannable doesn't define
a value for the given attribute, ignoring an Attr global default.

This behavior is tested with a call to scanOrDefault on all attributes.
Added missing coverage of scan for TAGS and NAME attributes.