fix #777 Offer way to do requests on original thread with subscribeOn
This commit adds a requestOnSeparateThread parameter to subscribeOn. Its
purpose is to make possible to NOT do the requests on the subscribeOn
worker when set to `false` for cases like when there is an eager or
blocking create as the source.

This case would previously deadlock because having the requests
would pile up behind the eager generator processing.