fix #789 Rename Mono.when to zip, supersede old Mono.zip(Iterable)
This commit renames Mono#when and Mono#whenDelayError to Mono#zip and
Mono#zipDelayError respectively. The previously existing zip functions
(with an iterable and a vararg) have been replaced in the process.

when signatures are deprecated. Related Kotlin extensions (and notably
`whenComplete`) are deprecated too as, unlike `when`, `zip` isn't a
keyword in Kotlin. All will be removed by 3.1.0.RELEASE (see #323).