Rename Mono.when to zip, supersede old Mono.zip(Iterable) (#794)
This commit renames Mono#when and Mono#whenDelayError to Mono#zip and
Mono#zipDelayError respectively, to the exception of `when` variants
that return a `Mono<Void>`.

The previously existing zip functions (with an iterable and a vararg)
have been replaced in the process.

Renamed `when` signatures are deprecated. Related Kotlin extensions (and
notably `whenComplete`) are deprecated too as, unlike `when`, `zip`
isn't a keyword in Kotlin. All will be removed by 3.1.0.RELEASE (#323).

Mono#and is also aliased as zipWith (both these variants will be kept).