First editing pass
I did a basic proofreading and added some details in the form of a section about threading and sample implementations of the various subscribe methods.

I put the new content in separate documetns and used include statements to insert them. That way, they can be reviewed and maintained separately and be easily rejected if need be.