fix #895 Safer scan of ACTUAL and PARENT
This commit makes scanning of ACTUAL and PARENT safer by introducing a
safe converter in the Attr class. By default, no converter is defined
and the old behavior of force-casting is still used.

For the two attributes above however, the value from `scanUnsafe` goes
through `Scannable#from` first when calling `scan` or `scanOrDefault`.
That way, even though a lot of operators blindly return something that
is Scannable _in most case but not necessarily_ (e.g. a Publisher),
the outer scan will not fail.