fix #922 Flux.fromStream(Supplier<Stream>) + auto-close Stream
This commit allows for lazy supplying of `Stream` so that multiple
subscriptions can work (by introducing a `Supplier<Stream>` variant).

It also modifies the operator so that `Stream` are automatically closed
when the subscription is cancelled or the sequence terminates. Although
most Streams don't need to be closed, it should not add overhead as the
close operation is no-op. For Streams that DO need to be closed however,
like IO Streams in the Files class, this removes the concern from the
user (same as try-with-resource would).