Improve documentation of Flux.merge
Avoid talking about "inner sources" or "inner publishers" when there is only
one collection of sources in sight. Reserve the use of the adjective "inner"
for variants of Flux.merge which deal with publishers of publishers.

Also, use the term "sources" instead of "publishers" consistently across the
related variants of Flux.merge.