fix #916 Process fatal exceptions in Schedulers.handleError, don't throw
The fatal exceptions will probably be thrown down the line and they are
still considered irrecoverable, but when they happen within a Thread
from a Scheduler, they should be passed to the handleError hook
as well as to the Thread's uncaughtExceptionHandler if any.

Such a sequence may hang, as the test demonstrates, but the fatal error
has a chance to be logged.