fix #1048 Detect Context trigger object in repeatWhen/retryWhen
This commit allows repeatWhen and retryWhen to detect that their trigger
for re-subscription was a `Context`, in which case the content of said
Context is added (`Context#addAll`) to the operator's own `Context`.

The coordinator `Subscriber` of each operator caches its `Context` for
the purpose of `actualContext()` and in order to make this enrichment
possible.