fix #1054 Distinct allows for arbitrary store and duplicate check
This commit removes the assumption that the distinct operator performs
its check via storing in a `Collection` and the `add` method. Instead,
it lets the user define an arbitrary store and a `BiPredicate` that
takes that store and the extracted key in order to determine if a value
has been encountered before.

A cleanup consumer is also used on termination, although like in the
former `Collection#clear`-based implementation it is assumed to not
raise an Exception.