fix #1080 Only terminate ExecutorScheduler if underlying is shut down
This commit prevents the termination of an `ExecutorScheduler` upon any
exception caught during submit. Such an exception could be transient,
coming from the task itself or even an Executor backed by a temporarily
full queue.

There is still a case where the scheduler can be terminated by a failure
in submit: since `ExecutorScheduler` _can_ be created with a
`ExecutorService`, the scheduler will detect that and check if the
`isShutdown()` status is true, and _only_ in this case would the
Scheduler be terminated.