fix #1077 Allow naming of a whole StepVerifier scenario through options
This commit allows a user to name a given `StepVerifier` by providing
a `StepVerifierOptions` that has a `scenarioName(String)` set.

The scenario name will appear in all StepVerifier-produced exceptions
as a prefix wrapped in square brackets, followed by a single space.

Methods of StepVerifier that execute user code which can throw 3rd party
Exceptions will continue propagating these as-is, without the scenario
name prefix (eg. assertNext).

Additionally, the formatting of error messages has been externalized in
a package-protected ErrorFormatter utility class, centralizing the
assembly of error messages (and the prefixing with the scenario name).